# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "ModelCandidateCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "context_window_tokens": {
            "exclusiveMinimum": 0,
            "title": "Context Window Tokens",
            "type": "integer"
          },
          "embeddings_support": {
            "default": false,
            "title": "Embeddings Support",
            "type": "boolean"
          },
          "input_modalities": {
            "items": {
              "type": "string"
            },
            "title": "Input Modalities",
            "type": "array"
          },
          "logical_model_id": {
            "title": "Logical Model Id",
            "type": "string"
          },
          "max_output_tokens": {
            "minimum": 0,
            "title": "Max Output Tokens",
            "type": "integer"
          },
          "native_features": {
            "items": {
              "type": "string"
            },
            "title": "Native Features",
            "type": "array"
          },
          "output_modalities": {
            "items": {
              "type": "string"
            },
            "title": "Output Modalities",
            "type": "array"
          },
          "pricing": {
            "additionalProperties": true,
            "title": "Pricing",
            "type": "object"
          },
          "priority": {
            "default": 100,
            "title": "Priority",
            "type": "integer"
          },
          "provider_connection_id": {
            "title": "Provider Connection Id",
            "type": "string"
          },
          "regions": {
            "items": {
              "type": "string"
            },
            "title": "Regions",
            "type": "array"
          },
          "retention_class": {
            "default": "standard",
            "title": "Retention Class",
            "type": "string"
          },
          "schema_support": {
            "default": "none",
            "enum": [
              "none",
              "json_mode",
              "best_effort_schema",
              "strict_json_schema"
            ],
            "title": "Schema Support",
            "type": "string"
          },
          "status": {
            "default": "active",
            "enum": [
              "active",
              "disabled"
            ],
            "title": "Status",
            "type": "string"
          },
          "streaming_support": {
            "default": false,
            "title": "Streaming Support",
            "type": "boolean"
          },
          "tool_modes": {
            "items": {
              "type": "string"
            },
            "title": "Tool Modes",
            "type": "array"
          },
          "unsupported_parameters": {
            "items": {
              "type": "string"
            },
            "title": "Unsupported Parameters",
            "type": "array"
          },
          "upstream_model_id": {
            "title": "Upstream Model Id",
            "type": "string"
          },
          "weight": {
            "default": 1,
            "exclusiveMinimum": 0,
            "title": "Weight",
            "type": "integer"
          }
        },
        "required": [
          "logical_model_id",
          "provider_connection_id",
          "upstream_model_id",
          "context_window_tokens",
          "max_output_tokens"
        ],
        "title": "ModelCandidateCreateRequest",
        "type": "object"
      },
      "ModelCandidateList": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ModelCandidateRecord"
            },
            "title": "Data",
            "type": "array"
          },
          "object": {
            "const": "list",
            "title": "Object",
            "type": "string"
          }
        },
        "required": [
          "object",
          "data"
        ],
        "title": "ModelCandidateList",
        "type": "object"
      },
      "ModelCandidateMutation": {
        "additionalProperties": false,
        "properties": {
          "candidate": {
            "$ref": "#/components/schemas/ModelCandidateRecord"
          },
          "configuration_version": {
            "$ref": "#/components/schemas/ConfigurationVersionSummary"
          }
        },
        "required": [
          "candidate",
          "configuration_version"
        ],
        "title": "ModelCandidateMutation",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "AdminBearerAuth": {
        "scheme": "bearer",
        "type": "http"
      },
      "ServiceBearerAuth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Unified API for routed model inference and gateway administration.",
    "title": "MixAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/admin/v1/candidates": {
      "get": {
        "operationId": "listModelCandidates",
        "parameters": [
          {
            "in": "query",
            "name": "logical_model_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Logical Model Id"
            }
          },
          {
            "description": "Optional caller-provided request identifier.",
            "in": "header",
            "name": "X-Request-ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional distributed trace context.",
            "in": "header",
            "name": "traceparent",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelCandidateList"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          },
          "402": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          }
        },
        "security": [
          {
            "AdminBearerAuth": []
          }
        ],
        "summary": "List Candidates",
        "tags": [
          "Admin"
        ]
      },
      "post": {
        "operationId": "createModelCandidate",
        "parameters": [
          {
            "description": "Optional caller-provided request identifier.",
            "in": "header",
            "name": "X-Request-ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional distributed trace context.",
            "in": "header",
            "name": "traceparent",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelCandidateCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelCandidateMutation"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          },
          "402": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          }
        },
        "security": [
          {
            "AdminBearerAuth": []
          }
        ],
        "summary": "Create Candidate",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/v1/candidates/{candidate_id}": {
      "delete": {
        "operationId": "deleteModelCandidate",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          },
          {
            "description": "Optional caller-provided request identifier.",
            "in": "header",
            "name": "X-Request-ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional distributed trace context.",
            "in": "header",
            "name": "traceparent",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelCandidateMutation"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          },
          "402": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          }
        },
        "security": [
          {
            "AdminBearerAuth": []
          }
        ],
        "summary": "Delete Candidate",
        "tags": [
          "Admin"
        ]
      },
      "get": {
        "operationId": "getModelCandidate",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          },
          {
            "description": "Optional caller-provided request identifier.",
            "in": "header",
            "name": "X-Request-ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional distributed trace context.",
            "in": "header",
            "name": "traceparent",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelCandidateRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          },
          "402": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          }
        },
        "security": [
          {
            "AdminBearerAuth": []
          }
        ],
        "summary": "Get Candidate",
        "tags": [
          "Admin"
        ]
      },
      "patch": {
        "operationId": "updateModelCandidate",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          },
          {
            "description": "Optional caller-provided request identifier.",
            "in": "header",
            "name": "X-Request-ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional distributed trace context.",
            "in": "header",
            "name": "traceparent",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelCandidateUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelCandidateMutation"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          },
          "402": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized MixAPI error"
          }
        },
        "security": [
          {
            "AdminBearerAuth": []
          }
        ],
        "summary": "Update Candidate",
        "tags": [
          "Admin"
        ]
      }
    },
...
  }
}